6/17/2025 Updated HTML, CSS and JavaScript.
Adjested HTML and CSS to be more responsive and made
the site look little better.

completed js
Script makes an array that hold all the book objs
made a class that is used to create books
made a function that handles making and updating the books
and adds a uuid, wanted to make my own uuid instead of usig
crypto.randomUUID()

when hitting submit a new book is made and added to the end
of the array. as well as a update button attached
to each book. when you click the update button it changes
the read status.

submit checks in place to make sure numbers arnt
0 or negatives and the title and auther is filled

no checks for dupe books on uuid's.